output opentelemetry: move to using the in k6 code
