+++ internal/output/opentelemetry/attribute.go
