+++ internal/output/opentelemetry/config.go
