+++ internal/output/opentelemetry/config_test.go
