+++ internal/output/opentelemetry/registry.go
